fastapi-minio/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
minio==7.1.17
python-dotenv==1.0.0
